Refactoring.

+++ service/src/test/kotlin/net/corda/examples/oracle/service/PrimesServiceTests.kt
@@ -19,7 +19,7 @@ class PrimesServiceTests : TestDependencyInjectionBase() {
 
     @Test
     fun `successful query`() {
-        assertEquals(104729, oracle.query(10000).toInt())
+        assertEquals(104729, oracle.query(10000))
     }
 
     @Test
